List Packages

+++ server/src/config/database.js
@@ -105,7 +105,7 @@ export async function initDatabase(conn) {
     promiseQuery(`CREATE TABLE IF NOT EXISTS ${PREFIX}package (
       id VARCHAR(50) NOT NULL PRIMARY KEY,
       user_id VARCHAR(100) NOT NULL,
-      price FLOAT(10, 2) UNSIGNED NOT NULL,
+      price FLOAT(20, 2) UNSIGNED NOT NULL,
       currency VARCHAR(10) NOT NULL,
       duration INT(10) UNSIGNED NOT NULL,
       register_date DATE NOT NULL,
